pyproject.toml
[project]
name = "tourwings-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql==0.2.0",
    "alembic==1.13.2",
    "cryptography>=45.0.6",
    "email-validator==2.2.0",
    "fastapi==0.112.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings==2.4.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "sqlmodel==0.0.22",
    "uvicorn[standard]==0.30.1",
]
